Fix captcha retrieval after validation errors
Fixes OpenSourcePolitics/decidim-module-question_captcha#9
See https://github.com/matthutchinson/acts_as_textcaptcha/issues/29

It is important to always add the textcaptcha_key and textcaptcha_answer
inputs, even when the captcha question is not being shown.
acts_as_textcaptcha offers a helper `textcaptcha_fields` for this, which
was used in this module, but inside an `if` statement. This led to wrong
behaviour when validation errors occurred on the registration form.